+++ .app/lib/modules/assets/index.js
@@ -1,11 +1,11 @@
-module.exports = {
-  _transformParser: require("./parser"),
+import { transformParser } from "./parser.js";
 
+export const assetsModule = {
   /**
    * Sets up the module.
    * @param {import("@11ty/eleventy").UserConfig} config
    */
   setup(config) {
-    config.addTransform(`assets-transform-parser`, this._transformParser);
+    config.addTransform(`assets-transform-parser`, transformParser);
   },
 };