+++ .app/lib/modules/dynamic-content/index.js
@@ -1,11 +1,14 @@
-module.exports = {
-  QueryRunner: require("./query-runner"),
-  TreeGenerator: require("./tree-generator"),
-  queryFilter: require("./query.filter"),
-  renderAsListFilter: require("./render-as-list.filter"),
+import { QueryRunner } from "./query-runner.js";
+import { TreeGenerator } from "./tree-generator.js";
+import { queryFilter } from "./query.filter.js";
+import { renderAsListFilter } from "./render-as-list.filter.js";
+
+export const dynamicContentModule = {
+  QueryRunner,
+  TreeGenerator,
 
   setup(config) {
-    config.addFilter("query", this.queryFilter(config));
-    config.addFilter("renderAsList", this.renderAsListFilter(config));
+    config.addFilter("query", queryFilter(config));
+    config.addFilter("renderAsList", renderAsListFilter(config));
   },
 };