+++ .app/lib/modules/notes/_notes.collection.js
@@ -1,4 +1,4 @@
-module.exports = () => (collectionApi) => {
+export const _notesCollection = () => (collectionApi) => {
   return collectionApi.getFilteredByGlob("../**/*.md").sort((a, b) => {
     const nameA = a.data.title || a.fileSlug;
     const nameB = b.data.title || b.fileSlug;