+++ .app/lib/modules/notes/copy-code.md-plugin.js
@@ -1,7 +1,7 @@
-const html = require("nanohtml");
-const raw = require("nanohtml/raw");
+import html from "nanohtml";
+import raw from "nanohtml/raw.js";
 
-module.exports = (md) => {
+export const copyCodeMarkdownPlugin = (md) => {
   md.renderer.rules.code_block = renderCode(md.renderer.rules.code_block);
   md.renderer.rules.fence = renderCode(md.renderer.rules.fence);