+++ .app/lib/modules/search/search-index.js
@@ -1,5 +1,6 @@
-module.exports = async (outputDir) => {
-  const pagefind = await import("pagefind");
+import * as pagefind from "pagefind";
+
+export const createIndex = async (outputDir) => {
   const { index } = await pagefind.createIndex();
 
   await index.addDirectory({ path: outputDir });