+++ .app/lib/shared/slugify-path.filter.js
@@ -3,7 +3,7 @@
  * @param {import("@11ty/eleventy").UserConfig} eleventyConfig
  * @returns {((path: string) => string)} The filter function.
  */
-module.exports = (eleventyConfig) => (path) => {
+export const slugifyPathFilter = (eleventyConfig) => (path) => {
   const separator = "/";
   const slugify = eleventyConfig.getFilter("slugify");